chore(broker-core): move time and variable record to protocol impl module
