+++ broker-core/src/test/java/io/zeebe/broker/workflow/timer/TimerStreamProcessorTest.java
@@ -23,10 +23,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.zeebe.broker.logstreams.processor.TypedRecord;
 import io.zeebe.broker.util.StreamProcessorControl;
 import io.zeebe.broker.util.StreamProcessorRule;
-import io.zeebe.broker.workflow.data.TimerRecord;
 import io.zeebe.broker.workflow.processor.WorkflowInstanceStreamProcessorRule;
 import io.zeebe.model.bpmn.Bpmn;
 import io.zeebe.protocol.clientapi.RejectionType;
+import io.zeebe.protocol.impl.record.value.timer.TimerRecord;
 import io.zeebe.protocol.intent.TimerIntent;
 import org.junit.Before;
 import org.junit.Rule;
